+++ lib/enhanced_composited_transform.dart
@@ -1,0 +1,6 @@
+export 'src/enhanced_composited_transform/anchor.dart'
+    show EnhancedCompositedTransformAnchor, EnhancedCompositedTransformAligned;
+export 'src/enhanced_composited_transform/flutter_src/rendering_layer.dart'
+    show EnhancedLayerLink;
+export 'src/enhanced_composited_transform/flutter_src/widgets_basic.dart'
+    show EnhancedCompositedTransformTarget, EnhancedCompositedTransformFollower;