+++ lib/src/enhanced_composited_transform/flutter_src/rendering_layer.dart
@@ -10,19 +10,17 @@ import 'package:flutter/rendering.dart';
 import 'package:flutter/scheduler.dart';
 import 'package:vector_math/vector_math_64.dart';
 
-import 'rendering_proxy_box.dart';
-
 /// @nodoc
-class CustomLayerLink {
-  /// The [CustomLeaderLayer] connected to this link.
-  CustomLeaderLayer? get leader => _leader;
-  CustomLeaderLayer? _leader;
+class EnhancedLayerLink {
+  /// The [EnhancedLeaderLayer] connected to this link.
+  EnhancedLeaderLayer? get leader => _leader;
+  EnhancedLeaderLayer? _leader;
 
-  void _registerLeader(CustomLeaderLayer leader) {
+  void _registerLeader(EnhancedLeaderLayer leader) {
     assert(_leader != leader);
     assert(() {
       if (_leader != null) {
-        _debugPreviousLeaders ??= <CustomLeaderLayer>{};
+        _debugPreviousLeaders ??= <EnhancedLeaderLayer>{};
         _debugScheduleLeadersCleanUpCheck();
         return _debugPreviousLeaders!.add(_leader!);
       }
@@ -31,7 +29,7 @@ class CustomLayerLink {
     _leader = leader;
   }
 
-  void _unregisterLeader(CustomLeaderLayer leader) {
+  void _unregisterLeader(EnhancedLeaderLayer leader) {
     if (_leader == leader) {
       _leader = null;
     } else {
@@ -40,7 +38,7 @@ class CustomLayerLink {
   }
 
   /// @nodoc
-  Set<CustomLeaderLayer>? _debugPreviousLeaders;
+  Set<EnhancedLeaderLayer>? _debugPreviousLeaders;
   bool _debugLeaderCheckScheduled = false;
 
   /// @nodoc
@@ -66,26 +64,26 @@ class CustomLayerLink {
       '${describeIdentity(this)}(${_leader != null ? "<linked>" : "<dangling>"})';
 }
 
-typedef PortalTheaterToLeaderOffset = Offset? Function();
+typedef TheaterRectRelativeToLeader = Rect Function();
 
 /// @nodoc
-class CustomLeaderLayer extends ContainerLayer {
+class EnhancedLeaderLayer extends ContainerLayer {
   /// @nodoc
-  CustomLeaderLayer({
-    required CustomLayerLink link,
+  EnhancedLeaderLayer({
+    required EnhancedLayerLink link,
     Offset offset = Offset.zero,
-    required PortalTheaterToLeaderOffset portalTheaterToLeaderOffset,
+    required TheaterRectRelativeToLeader theaterRectRelativeToLeader,
     required this.debugName,
   })  : assert(link != null),
         _link = link,
-        _portalTheaterToLeaderOffset = portalTheaterToLeaderOffset,
+        _theaterRectRelativeToLeader = theaterRectRelativeToLeader,
         _offset = offset;
 
   /// @nodoc
-  CustomLayerLink get link => _link;
-  CustomLayerLink _link;
+  EnhancedLayerLink get link => _link;
+  EnhancedLayerLink _link;
 
-  set link(CustomLayerLink value) {
+  set link(EnhancedLayerLink value) {
     assert(value != null);
     if (_link == value) {
       return;
@@ -113,16 +111,16 @@ class CustomLeaderLayer extends ContainerLayer {
   }
 
   /// @nodoc
-  PortalTheaterToLeaderOffset get portalTheaterToLeaderOffset =>
-      _portalTheaterToLeaderOffset;
-  PortalTheaterToLeaderOffset _portalTheaterToLeaderOffset;
+  TheaterRectRelativeToLeader get theaterRectRelativeToLeader =>
+      _theaterRectRelativeToLeader;
+  TheaterRectRelativeToLeader _theaterRectRelativeToLeader;
 
-  set portalTheaterToLeaderOffset(PortalTheaterToLeaderOffset value) {
+  set theaterRectRelativeToLeader(TheaterRectRelativeToLeader value) {
     assert(value != null);
-    if (value == _portalTheaterToLeaderOffset) {
+    if (value == _theaterRectRelativeToLeader) {
       return;
     }
-    _portalTheaterToLeaderOffset = value;
+    _theaterRectRelativeToLeader = value;
   }
 
   // NOTE MODIFIED add
@@ -172,8 +170,8 @@ class CustomLeaderLayer extends ContainerLayer {
     super.debugFillProperties(properties);
     properties.add(DiagnosticsProperty<Offset>('offset', offset));
     properties.add(DiagnosticsProperty(
-        'portalTheaterToLeaderOffset', portalTheaterToLeaderOffset));
-    properties.add(DiagnosticsProperty<CustomLayerLink>('link', link));
+        'theaterRectRelativeToLeader', theaterRectRelativeToLeader));
+    properties.add(DiagnosticsProperty<EnhancedLayerLink>('link', link));
     properties.add(DiagnosticsProperty('debugName', debugName));
   }
 }
@@ -191,13 +189,14 @@ class CustomLeaderLayer extends ContainerLayer {
 /// [FollowerLayer.unlinkedOffset] being [Offset.zero]).
 ///
 /// For documentation of undocumented code, see [FollowerLayer].
-class CustomFollowerLayer extends ContainerLayer {
+class EnhancedFollowerLayer extends ContainerLayer {
   // NOTE MODIFIED the comments
   /// Creates a follower layer.
   ///
   /// The [link] property must not be null.
-  CustomFollowerLayer({
+  EnhancedFollowerLayer({
     required this.link,
+    required this.showWhenUnlinked,
     // NOTE MODIFIED add [linkedOffsetCallback], remove several arguments like
     // [showWhenUnlinked], [unlinkedOffset], [linkedOffset]
     required this.linkedOffsetCallback,
@@ -205,7 +204,9 @@ class CustomFollowerLayer extends ContainerLayer {
     required this.debugName,
   });
 
-  CustomLayerLink link;
+  EnhancedLayerLink link;
+
+  bool? showWhenUnlinked;
 
   // NOTE MODIFIED added this field
   /// Callback that is called to compute the linked offset of the follower layer
@@ -227,9 +228,6 @@ class CustomFollowerLayer extends ContainerLayer {
   String? debugName;
 
   Offset? unlinkedOffset;
-
-  // NOTE MODIFIED similarly, make [showWhenUnlinked] a const for our needs.
-  static const showWhenUnlinked = CustomRenderFollowerLayer.showWhenUnlinked;
 
   Offset? _transformOffset(Offset localPosition) {
     if (_inverseDirty) {
@@ -253,7 +251,7 @@ class CustomFollowerLayer extends ContainerLayer {
       AnnotationResult<S> result, Offset localPosition,
       {required bool onlyFirst}) {
     if (link.leader == null) {
-      if (showWhenUnlinked) {
+      if (showWhenUnlinked!) {
         return super.findAnnotations(result, localPosition - unlinkedOffset!,
             onlyFirst: onlyFirst);
       }
@@ -414,7 +412,8 @@ class CustomFollowerLayer extends ContainerLayer {
 
   @override
   void addToScene(ui.SceneBuilder builder) {
-    if (link.leader == null && !showWhenUnlinked) {
+    assert(showWhenUnlinked != null);
+    if (link.leader == null && !showWhenUnlinked!) {
       _lastTransform = null;
       _lastOffset = null;
       _inverseDirty = true;
@@ -464,7 +463,7 @@ class CustomFollowerLayer extends ContainerLayer {
   @override
   void debugFillProperties(DiagnosticPropertiesBuilder properties) {
     super.debugFillProperties(properties);
-    properties.add(DiagnosticsProperty<CustomLayerLink>('link', link));
+    properties.add(DiagnosticsProperty<EnhancedLayerLink>('link', link));
     properties.add(
         TransformProperty('transform', getLastTransform(), defaultValue: null));
     // NOTE MODIFIED