+++ lib/src/portal_target.dart
@@ -6,8 +6,8 @@ import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 
 import 'anchor.dart';
-import 'flutter_src/rendering_layer.dart';
-import 'flutter_src/widgets_basic.dart';
+import 'enhanced_composited_transform/flutter_src/rendering_layer.dart';
+import 'enhanced_composited_transform/flutter_src/widgets_basic.dart';
 import 'portal.dart';
 import 'portal_link.dart';
 import 'portal_target_theater.dart';
@@ -265,7 +265,7 @@ class PortalTarget extends StatefulWidget {
 }
 
 class _PortalTargetState extends State<PortalTarget> {
-  final _link = CustomLayerLink();
+  final _link = EnhancedLayerLink();
 
   @override
   Widget build(BuildContext context) {
@@ -294,9 +294,10 @@ class _PortalTargetState extends State<PortalTarget> {
 
     return Stack(
       children: <Widget>[
-        CustomCompositedTransformTarget(
+        EnhancedCompositedTransformTarget(
           link: _link,
-          portalLink: scope.portalLink,
+          // TODO improve it: currently a new instance of `theaterGetter` will trigger repaint etc
+          theaterGetter: () => scope.portalLink.theater,
           debugName: widget.debugName,
           child: widget.child,
         ),
@@ -310,9 +311,9 @@ class _PortalTargetState extends State<PortalTarget> {
                   portalLink: scope.portalLink,
                   anchor: widget.anchor,
                   targetSize: targetSize,
-                  portalFollower: CustomCompositedTransformFollower(
+                  portalFollower: EnhancedCompositedTransformFollower(
                     link: _link,
-                    portalLink: scope.portalLink,
+                    showWhenUnlinked: false,
                     anchor: widget.anchor,
                     targetSize: targetSize,
                     debugName: widget.debugName,